feat: Add shell support and enable fs commands
This enables shell support on the USB CDC ACM serial connection and
turns on the filesystem shell commands in order to be able to obtain the
logs without pulling the SD card out. Down the line other shell commands
will be added as well, for instance for ESC calibration or parameter
changes.